% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowerpower.R
\name{date_ranges}
\alias{date_ranges}
\title{Internal function to create data ranges (since max date span is 10 days)}
\usage{
date_ranges(from, to = now(), interval = days(9))
}
\description{
Internal function to create data ranges (since max date span is 10 days)
}

