% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowerpower.R
\name{flowerpower}
\alias{flowerpower}
\title{Create new API connection to Parrot's FlowerPower sensor}
\usage{
flowerpower(user, pass, client_id, client_secret)
}
\arguments{
\item{user}{username}

\item{pass}{password}

\item{client_id}{client identifier}

\item{client_secret}{client secret key}
}
\description{
Create new API connection to Parrot's FlowerPower sensor
}

