% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowerpower.R
\docType{data}
\name{flowerpower_factory}
\alias{flowerpower_factory}
\title{All internal methods are in R6, API funs in S3}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fe3bb88cf78> 
 - attr(*, "name")= chr "flowerpower_generator"
}}
\usage{
flowerpower_factory
}
\description{
All internal methods are in R6, API funs in S3
}
\keyword{datasets}

