% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowerpower.R
\name{get_fertilizer}
\alias{get_fertilizer}
\title{Get fertilizer samples.}
\usage{
get_fertilizer(obj, location)
}
\arguments{
\item{obj}{a flowerpower object}

\item{location}{a location identifier}
}
\description{
Get fertilizer samples.
}

