% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowerpower.R
\name{get_locations.flowerpower}
\alias{get_locations.flowerpower}
\title{Get all sensor locations}
\usage{
\method{get_locations}{flowerpower}(obj, location = NULL)
}
\arguments{
\item{obj}{a flowerpower object}

\item{location}{a location identifier}
}
\description{
Get all sensor locations
}

