% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowerpower.R
\name{get_locations}
\alias{get_locations}
\title{Get all sensor locations}
\usage{
get_locations(obj, location = NULL)
}
\arguments{
\item{obj}{a flowerpower object}

\item{location}{a location identifier}
}
\description{
Get all sensor locations
}

