% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowerpower.R
\name{get_samples.flowerpower}
\alias{get_samples.flowerpower}
\title{Get data samples.}
\usage{
\method{get_samples}{flowerpower}(obj, location)
}
\arguments{
\item{obj}{a flowerpower object}

\item{location}{a location identifier}
}
\description{
Get data samples.
}

