% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowerpower.R
\name{get_samples}
\alias{get_samples}
\title{Get data samples.}
\usage{
get_samples(obj, location)
}
\arguments{
\item{obj}{a flowerpower object}

\item{location}{a location identifier}
}
\description{
Get data samples.
}

